Hacks to make JavaScript play nice, sort of
This needs a lot of love but I took care of most of the loading bugs,
for now. Dynamically loaded content (i.e. the radar chart and select
dropdown) was being loaded twice when navigating with the forward/back
browser buttons. I suspect something to do with Turbo/browser cache
clashing.