Fix kill buttons with multiple trainees
Discovered while implementing tests that the kill buttons were broken;
they would update the inputs but only the last trainee would update on
the server... most of the time. This one was a doozie; Stimulus wasn't
playing nice rapid-firing requests like that and the server didn't like
it either so I'm using JavaScript to re-render the chart when you hit a
kill button. Updating the inputs still uses the Turbo Stream.

refs #2