Remove mini-sidenav on mobile
It was a waste of space. Discovered a couple of sidenav-related bugs
while doing this so I put some tests in for mobile screens, with room to
grow.

The built-in Materialize classes were giving me grief when navigating
via the expanded sidenav on mobile, so I had to close the nav and call
the pageloads onclick with JavaScript.